* only remove build binaries on debian

+++ ffmpeg-build.sh
@@ -853,8 +853,10 @@ function build_deb {
 
 function build_clean {
 
-    # remove build
-    rm -r -f $BIN_DIR/*
+    if [ "$ENVIRONMENT" == "deb" ]
+    then
+        rm -r -f $BIN_DIR/*
+    fi
 
     #remove sources
     rm -r -f $SRC_DIR/*
